requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.13.0
numpy==2.3.5
openai==1.58.1
packaging==24.2
pandas==2.3.3
pillow==11.3.0
pip==25.3
protobuf==5.29.5
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.41.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.2
